+++ app/src/main/java/com/example/firstandroid/build_config/BuildConfig.java
@@ -1,4 +1,4 @@
-package com.example.firstandroid.BuildConfig;
+package com.example.firstandroid.build_config;
 
 public final class BuildConfig {
     public static final String BASE_URL = "http://10.0.2.2:8080/api/v1/";
