+++ app/src/main/java/com/example/firstandroid/presentation/add_bathroom/NewBathroomScreen.kt
@@ -1,4 +1,4 @@
-package com.example.firstandroid.Presentation.UpdateBathroom
+package com.example.firstandroid.presentation.add_bathroom
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -10,58 +10,45 @@ import androidx.compose.material3.Checkbox
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 
+
 @Composable
-fun UpdateBathroomDialog(
+fun AddBathroomDialog(
     onDismissRequest: () -> Unit,
-    viewModel: UpdateBathroomViewModel,
-    originalBathroomInfo: UpdateBathroomData
+    viewModel: NewBathroomViewModel,
 ) {
 
-    val state = viewModel.state
     val scope = rememberCoroutineScope()
-    var title by remember { mutableStateOf(originalBathroomInfo.title) }
-    var location by remember { mutableStateOf(originalBathroomInfo.location) }
-    var capacity by remember { mutableStateOf(originalBathroomInfo.capacity.toString()) }
-    var free by remember { mutableStateOf(originalBathroomInfo.free) }
-    var cost by remember { mutableStateOf(originalBathroomInfo.cost.toString()) }
-    var hours by remember { mutableStateOf(originalBathroomInfo.hours) }
+
     AlertDialog(
         onDismissRequest = { onDismissRequest() },
         text = {
             Column {
                 TextField(
-                    value = title,
+                    value = viewModel.state.title,
                     onValueChange = { newTitle ->
-                        title = newTitle
-                        viewModel.updateTitle(title)
+                        viewModel.updateTitle(newTitle)
                     },
                     label = { Text("Title") }
                 )
                 Spacer(modifier = Modifier.height(16.dp))
                 TextField(
-                    value = location,
+                    value = viewModel.state.location,
                     onValueChange = { newLocation ->
-                        location = newLocation
-                        viewModel.updateLocation(location)
+                        viewModel.updateLocation(newLocation)
                     },
                     label = { Text("Location") }
                 )
                 Spacer(modifier = Modifier.height(16.dp))
                 TextField(
-                    value = capacity,
+                    value = viewModel.state.capacity.toString(),
                     onValueChange = { newCapacity ->
-                        capacity = newCapacity
-                        viewModel.updateCapacity(capacity)
+                        viewModel.updateCapacity(newCapacity)
                     },
                     label = { Text("Capacity") }
                 )
@@ -69,45 +56,39 @@ fun UpdateBathroomDialog(
                 Row(verticalAlignment = Alignment.CenterVertically) {
                     Text("Free")
                     Checkbox(
-                        checked = free,
+                        checked = viewModel.state.free,
                         onCheckedChange = { isChecked ->
-                            free = isChecked
-                            viewModel.updateFree(free)
+                            viewModel.updateFree(isChecked)
                         }
                     )
                 }
                 Spacer(modifier = Modifier.height(16.dp))
                 TextField(
-                    value = cost,
+                    value = viewModel.state.cost.toString(),
                     onValueChange = { newCost ->
-                        cost = newCost
-                        viewModel.updateCost(cost)
+                        viewModel.updateCost(newCost)
                     },
                     label = { Text("Cost") }
                 )
                 Spacer(modifier = Modifier.height(16.dp))
                 TextField(
-                    value = hours,
+                    value = viewModel.state.hours,
                     onValueChange = { newHours ->
-                        hours = newHours
-                        viewModel.updateHours(hours)
+                        viewModel.updateHours(newHours)
                     },
                     label = { Text("Hours") }
                 )
-
             }
         },
         confirmButton = {
             Button(
                 onClick = {
                     scope.launch {
-                        state.id = originalBathroomInfo.id
-                        viewModel.updateBathroom(state)
-                        onDismissRequest()
+                        viewModel.addBathroom(viewModel.state)
                     }
                 }
             ) {
-                Text("Update")
+                Text("Add")
             }
         },
         dismissButton = {
