+++ app/src/main/java/com/example/firstandroid/presentation/add_bathroom/NewBathroomState.kt
@@ -1,9 +1,9 @@
-package com.example.firstandroid.Presentation.NewBathroom
+package com.example.firstandroid.presentation.add_bathroom
 
 import java.math.BigDecimal
 
 //declare data
-data class NewBathroomData(
+data class NewBathroomState(
     val title: String = "",
     val location: String = "",
     val capacity: Int = 0,
