+++ app/src/main/java/com/example/firstandroid/presentation/get_bathroom/GetBathroomState.kt
@@ -1,0 +1,20 @@
+package com.example.firstandroid.presentation.get_bathroom
+
+import java.math.BigDecimal
+
+data class GetBathroomData(
+    val id: Long? = 0L,
+    val title: String? = "",
+    val location: String? = "",
+    val capacity: Int? = 0,
+    val free: Boolean? = true,
+    val cost: BigDecimal? = BigDecimal.ZERO,
+    val hours: String? = "",
+    val createdAt: String? = "",
+    val modifiedAt: String? = "",
+)
+data class GetBathroomDataList(
+    val getBathroomList: List<GetBathroomData> = emptyList()
+)
+
+
